feat: release version 0.5.0 with major breaking changes
This major release introduces several breaking changes that simplify the process of coding autograms.

These changes make autograms a true python library, as opposed to a graph language with python-like syntax. It maintains the mapping between graphs and code.

In the old (0.1) autograms, python-like code is compiled into a graph and interpreted as a graph, in contrast the new (0.5) autograms interprets python code directly, but uses a special function decorator to allow features like:
    serializing the exact program state
    replying from anywhere in the program and pausing execution
    jumping to predefined locations (nodes) in the program

This enables behavior equivalent to autograms 0.1 while allowing python code be be run directly rather than compiled into graphical form, allowing much more flexibility with integrating other python code.

Autograms 0.5 still allows python code to be compiled into a graph for visualization (allowing loops, conditionals, and GOTOs (if any) to be visualized), as well as for a graph representation to be decompiled back into python code, however like with autograms 0.1 this may not support all python.

In the new version, each subgraph or function is designated by using the `@autograms_function()` decorator around a python function. This allows them to be easily integrated wth normal python code.

The reply and reply_instruction functions are similar to the chat_exact and chat nodes respectively in autograms 0.1. They allow you to reply from anywhere in the graph/program while saving and serializing the state.

The special `ADDRESS` argument add argument passed to an `@autograms_node` decorated function has similar functionality to the `name` field in autograms 0.1, and allows nodes to be addressed

Behavior equivalent to the original autograms transitions can either be accomplishes using the special TRANSITION function, or a combination of the multiple_choice function and GOTO function that allows you to jump to a node with an ADDRESS

The Autogram object is designed to wrap the root @autograms_function() function of your chatbot or agent

Libraries for manipulating prompts and memory are more accessible from within the code now, whereas these were hidden deep in the interpreter before.